default.ui: let the toolbars appear when needed
This was often requested by users so that I already implemented this in the alternative installer.
This furthermore avoids questions on the lyx-users list (we had several times questions about "Where is the math toolbar when I'm in math?" etc.)
(This toolbar behavior is also used e.g. by OpenOffice or Word.)

As this topic now arises again with the macro toolbar
http://www.mail-archive.com/lyx-devel@lists.lyx.org/msg136267.html
and nobody objected, I implemented this now.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@23370 a592a061-630c-0410-9148-cb99ea01b6c8