Forgot this bit.
The only remaining problem here is that, if you edit the TEXT that goes with
an InsetBibitem, then the information in the citation will not get updated
until something else invalidates the cache.  The only way to fix this, given
the existing system, would seem to be to check, every time anything gets 
edited, whether we're in a Bibliography environment or not, and if so to
invalidate the cache. I'm not sure that is worth it. Really, it points to a
more general issue with InsetBibitem, IMHO: It ought to be a kind of combined
inset, a hybrid of sorts between an InsetCollapsable and an InsetCommand. But
that would be for 1.7.x at least.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@25138 a592a061-630c-0410-9148-cb99ea01b6c8