This patch allows for multiple LFUNs to be bound to a single key, via a new command-alternatives function, to be used like so: 	\bind "Tab" "command-alternatives" "completion-accept;cell-forward" The first of the metioned functions that is enabled with be called.
The code is pretty trivial and just stolen from command-sequence.

The binding mentioned is more or less the point of the patch, but we still need to decide exactly what the bindings should be. There have been various threads concerning this. You can use the above for testing. Add it to site.bind, replacing the existing binding.

Jose: Once the bindings have been sorted out, that will complete the work on this we said we wanted to do for rc3.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@26518 a592a061-630c-0410-9148-cb99ea01b6c8