Fix bug 5330, round 1.
The bug is that older layout files with lines like:
	Input amsmaths.inc
fail in 1.6, since that file no longer exists. This commit adds some 
layout2layout that will scan for such includes and add lines like
	UseModule theorems-ams
corresponding to the input lines. The downside to this commit is that
UseModule only becomes available on 25 July 2008, whereas the files
were removed in January.

My intention is to try to deal with these intermediate cases as well.
But I'm not sure that will ultimately work very well, so this will do
until I can figure that out.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@26800 a592a061-630c-0410-9148-cb99ea01b6c8