Fix bug 5699 (crash when using outliner in child document while the master is not opened).
The problem was that, if the master is not opened, no respective guiDelegate exists, and the functions that update the toc do not trigger. The fix is to update the child's toc backend and toc items directly in such cases.

All credits (for both detecting and fixing the problem) go to Vincent, who is going to submit the (slightly different) fix to trunk.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/branches/BRANCH_1_6_X@28527 a592a061-630c-0410-9148-cb99ea01b6c8