Fix bug 4346: http://bugzilla.lyx.org/show_bug.cgi?id=4346. Synchronizing insets asserts with two views open
It changes the autoOpen_ member of InsetCollapsable into a map, such that an autoOpen value can be specified for each bufferView. Now, the assertion is avoided and insetCollapsable can be open in one bufferview and be closed in the other in very special cases that the cursor end up in a closed inset.

Compare with the MathMacro::editing_ member. 

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/branches/BRANCH_1_6_X@28587 a592a061-630c-0410-9148-cb99ea01b6c8