branch: Fix bug 5088: Incorrect cursor movement http://bugzilla.lyx.org/show_bug.cgi?id=5088
After executing DEPM to remove a double space, the position of the cursor was only 'fixed' if the cursor remained in the same par. This patch corrects this
to fix the position of the cursor also when the cursor is in a nested inset which is in the same par. This is needed because otherwise the cursor would not
point to the inset but one pos after it.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/branches/BRANCH_1_6_X@28780 a592a061-630c-0410-9148-cb99ea01b6c8