New literate programming modules.
These modules allow to use Sweave or Noweb with any LyX textclass.
They will eventually allow to get rid of literate-* classes.

The sweave modules will benefit in trunk from a forthcoming patch which allows 
to select the output format for a literate OutputType. This will allow to
support Sweave out-of-the-box.

For branch, this pair of modules is IMO a candidate for LyX 1.6.4 (no need to
hurry).

Thanks to Gregor Gorjanc for his work on this subject. There is
undoubtlessly some additional work that has to be done here.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@29657 a592a061-630c-0410-9148-cb99ea01b6c8