backport rev. 29324:
Timestamp: 
04/19/09 13:57:52
Author: 
spitz 
Message: 
Implement a warning dialog with a "Do not show this warning again!" checkbox.
Since Qt does not provide this, we have to roll our own dialog.
Hidden warnings are stored in sessions.

The dialog is currently only used in the case where an included file
has a different textclass then the master (bug 3218).

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/branches/BRANCH_1_6_X@29941 a592a061-630c-0410-9148-cb99ea01b6c8