branch: Fix bug #4952: Multiple cells pasting fails if cells are selected.
When the clipboard contains a tabular selection, dirty_tabular_stack_ is true. However, it is set to false when the user clears a selection (without copying). After this (mostly accidental) action, pasting the tabular contents goes wrong.

So, we only clear the dirty_tabular_stack_ flag when a 'real' copy to clipboard has been made.

see r29717, r29718, r29719



git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/branches/BRANCH_1_6_X@30499 a592a061-630c-0410-9148-cb99ea01b6c8