Fix bug #6108: Crash on menu key
The wrong context menu is being expanded, because the location where the context menu is requested is computed wrongly by Qt. 

Actually, the problem seemed to be in InputMethodQuery?. For some reason, the box of the cursor is shifted right under the cursor box.

The menu key uses InputMethodQuery? to locate the context menu, just as he Japanese input method locates the box with possibilities.

1.6.x shows the same behaviour, but in that case it doesn't crash because the spellchecker entry does not exist there, but there might be another case in which it will crash.




git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@30797 a592a061-630c-0410-9148-cb99ea01b6c8