Fix bug 5946: Infinite loop when opening file with two files opened with the same name
This simply changes the appearance of the display path (in the tab header) in one detail: if the file suffix is not "lyx" (but "lyx15" etc.), the file is displayed with the extension. This fixes the problem (which is due to identical display paths), and it adds, IMHO, useful information to the GUI.

Patch reviewed by vfr.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/branches/BRANCH_1_6_X@30978 a592a061-630c-0410-9148-cb99ea01b6c8