Mark a buffer as clean when the user chose to discard the changes. This is a patch from Richard to make his emergency cleanup procedure in Buffer::~Buffer work.
Besides that, this can also been seen in 1.6.3. If you have a master and a dirty child, and the dirty child is the second tab, you'll see that when closing LyX, LyX will ask twice whether you want to save the file. Once if the file itself is closed, a second time if the master is closing all his children.

See also the previous commit that treats the case in which the tabs were switched around.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@31074 a592a061-630c-0410-9148-cb99ea01b6c8