If we are in a close_event all children will be closed in some time, so no need to do it here. This will ensure that the children end up in the session file in the correct order. If we close the master buffer, we can close or release the child buffers here too.
We don't want a master buffer to release all its children immediately because that would close dirty childs without saving and would make it impossible to store the childs in a proper way in the session file. 

As, we close all children of a buffer together with the master itself, it is also no longer necessary to do it automatically. 

I'll leave it commented out for a while in case someone objects to it.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@31130 a592a061-630c-0410-9148-cb99ea01b6c8