This inner tag thing turns out to be a bad idea, since so few things are allowed to go inside <span>. Probably we don't need it anyway.
The User's Guide now exports as valid XHTML 1.1, according to the XHTML 
validator at w3.org. Of course, in the present state of the code, a lot
of insets aren't being exported at all, since they've been disabled. So
there is work to do. But this is a serious start now.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@32108 a592a061-630c-0410-9148-cb99ea01b6c8