branch: Fix bugs - #6161: Wrong cursor position in RTL text - #6165: Assertion when clicking on left side of inset in RTL text
If the first character is a separator, we are in RTL text. This character will not be painted on screen and thus we should not count it and skip to the next.

This bug pretty much destroyed cursor placement in RTL text on a row with large gaps.

see r32116.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/branches/BRANCH_1_6_X@32121 a592a061-630c-0410-9148-cb99ea01b6c8