Restore XHTML output for InsetGraphics.
To make this really good, we'd want to be able to rotate, crop, etc, the
images, as well as convert them to suitable output formats. This should
not be too difficult, since we do this sort of thing for display,
anyway. But I don't myself know how to do it. Anyone?


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@32295 a592a061-630c-0410-9148-cb99ea01b6c8