branch: Fix bug #6365: tex2lyx uses wrong encoding in the preamble.
Everything is ok while reading the file, but when the lyxfile is written, only the first byte of a double-byte character is written to the file. This is because t.character() is sort of defined as cs_[0] in which cs_ is a docstring. This thus seems to fail for multi-byte characters.

see r32441.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/branches/BRANCH_1_6_X@32442 a592a061-630c-0410-9148-cb99ea01b6c8