branch: Refresh the contents of some dialogs when the document is read-only.
Update the view also when the document is read-only.

see r32694.

Update the GuiInfo dialog also when the document is read-only. A dialog that does not dispatch a single LFUN (see getLfun()) but is buffer dependent (see isBufferDependent()) should implement the canApply function (if I'm correct).

see r32695.

Update the GuiLabel view when the document is read-only. Also disable the text field.

see r32696.

git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/branches/BRANCH_1_6_X@32802 a592a061-630c-0410-9148-cb99ea01b6c8