Spellcheck related bugfix from John McCabe-Dansted:
PrefSpellchecker disables the "Spellchecker Engine:" selector if
theSpellChecker() reports that no spell checking engine is selected.
This means that if we have not already selected a spell checking
engine we cannot select one now. I think this problem can be
reproduced by first compiling and running LyX without spellchecking
support and then compiling and running LyX with spellchecking support.

In any case it is clear that we should be able to select a
spellchecking engine if one is available, so I propose the attached
patch. This patch fixed the problem on my machine.


git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@33669 a592a061-630c-0410-9148-cb99ea01b6c8