This is a dummy commit for posting the real changelog for the the last two commits. Sorry guys, I should have really gone to sleep...
r34320
In mathed, preserve the color in a proper way, such as to not interfere
with the colors normally used to represent math objects.
When changing the color of a chunk of text containing math formulas,
now the formulas are rendered on screen in the chosen color, too.

r34321
Assure that constructs such as \mathbf{a\ensuremath{b}} are correctly
rendered on screen.



git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@34322 a592a061-630c-0410-9148-cb99ea01b6c8