Initialize some LyX colors from Qt. Currently what is done is plain text and selection (both font and background).
I am not sure that there is something else that we can get from QPalette. Maybe
cursor color, if someone tells me how to get it.

Only tested on mac. I can revert if this causes problems on linux or windows.

I have to check too that it still works in non-gui mode.



git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@35060 a592a061-630c-0410-9148-cb99ea01b6c8