A go at creating a prefs2prefs framework. There's a main file, prefs2prefs.py, and two subsidiary files, prefs2prefs_lfuns.py and prefs2prefs_prefs.py. I've organized it this way because, in many ways, these are  the same task. It's very line-by-line, unlike lyx2lyx and layout2layout, where things can be more "global". So we read the file, line by line, and give a bunch of converter functions a chance to see if they want to modify that line.
The converter functions are all in the subsidiary files. (Only the lfun one has anything in it now.) They 
take a line as argument and return a list: (Bool, NewLine), where the Bool says if we've modified anything 
and the NewLine is the new line, if so.

The format of the existing files is format 0, and we'll introduce new format numbers as we proceed, 
just as with layout2layout. So the conversion from format 0 to format 1 will be huge; others will 
generally be simple.



git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@35426 a592a061-630c-0410-9148-cb99ea01b6c8